+++ test/webapi/ows/coverages/test_routes.py
@@ -21,17 +21,19 @@
 
 from ...helpers import RoutesTestCase
 
+from xcube.webapi.ows.coverages.routes import PATH_PREFIX
+
 
 class CoveragesRoutesTest(RoutesTestCase):
     def test_fetch_coverage(self):
         response = self.fetch(
-            '/catalog/collections/demo/coverage?f=application/json'
+            PATH_PREFIX + '/collections/demo/coverage?f=application/json'
         )
         self.assertResponseOK(response)
 
     def test_fetch_coverage_html(self):
         response = self.fetch(
-            '/catalog/collections/demo/coverage',
+            PATH_PREFIX + '/collections/demo/coverage',
             headers=dict(
                 Accept='text/nonexistent,application/json;q=0.9,text/html;q=1.0'
             ),
@@ -41,23 +43,31 @@ class CoveragesRoutesTest(RoutesTestCase):
 
     def test_fetch_coverage_wrong_media_type(self):
         response = self.fetch(
-            '/catalog/collections/demo/coverage',
+            PATH_PREFIX + '/collections/demo/coverage',
             headers=dict(Accept='text/nonexistent'),
         )
         self.assertEqual(response.status, 415)
 
     def test_fetch_domainset(self):
-        response = self.fetch('/catalog/collections/demo/coverage/domainset')
+        response = self.fetch(
+            PATH_PREFIX + '/collections/demo/coverage/domainset'
+        )
         self.assertResponseOK(response)
 
     def test_fetch_rangetype(self):
-        response = self.fetch('/catalog/collections/demo/coverage/rangetype')
+        response = self.fetch(
+            PATH_PREFIX + '/collections/demo/coverage/rangetype'
+        )
         self.assertResponseOK(response)
 
     def test_fetch_metadata(self):
-        response = self.fetch('/catalog/collections/demo/coverage/metadata')
+        response = self.fetch(
+            PATH_PREFIX + '/collections/demo/coverage/metadata'
+        )
         self.assertResponseOK(response)
 
     def test_fetch_rangeset(self):
-        response = self.fetch('/catalog/collections/demo/coverage/rangeset')
+        response = self.fetch(
+            PATH_PREFIX + '/collections/demo/coverage/rangeset'
+        )
         self.assertEqual(response.status, 501)