+++ xcube/webapi/ows/stac/config.py
@@ -26,9 +26,19 @@ DEFAULT_CATALOG_ID = "xcube-server"
 DEFAULT_CATALOG_TITLE = "xcube Server"
 DEFAULT_CATALOG_DESCRIPTION = "Catalog of datasets served by xcube."
 
+# ID, name, and description for the unified collection containing a feature
+# for each datacube
 DEFAULT_COLLECTION_ID = "datacubes"
 DEFAULT_COLLECTION_TITLE = "Data cubes"
 DEFAULT_COLLECTION_DESCRIPTION = "a collection of xcube datasets"
+
+# As well as the unified collection, there's an individual collection for
+# each datacube, with the same name as that datacube, and containing a single
+# feature representing that datacube. This is the name of that single feature.
+DEFAULT_FEATURE_ID = 'datacube'
+
+# Prefix for STAC and OGC endpoints
+PATH_PREFIX = '/ogc'
 
 COLLECTION_SCHEMA = JsonObjectSchema(
     properties=dict(